+++ go/cwc/io/nullio.go
@@ -15,7 +15,7 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
-package cwc
+package io
 
 // Null I/O type
 // useful for testing, doesn't aim to do anything
