+++ go/cwc/io/pigpio.go
@@ -1,5 +1,4 @@
 /*
-Copyright (C) 2019 Andrew Amos
 Copyright (C) 2019 Graeme Sutherland, Nodestone Limited
 
 
@@ -16,7 +15,7 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
-package cwc
+package io
 
 import (
 	"github.com/golang/glog"
@@ -30,29 +29,28 @@ import (
  */
 
 // PWM settings
-// const OnDutyCycle = uint32(1)
-// const PWMCycleLength = uint32(32)
+const OnDutyCycle = uint32(1)
+const PWMCycleLength = uint32(32)
 
-type Keyer struct {
-	config   *Config
-	output   rpio.Pin
-	dahInput rpio.Pin
-	ditInput rpio.Pin
-	pwm      bool
-	pwmOut   rpio.Pin
-	status   rpio.Pin
+type PiGPIO struct {
+	config *Config
+	output rpio.Pin
+	input  rpio.Pin
+	pwm    bool
+	pwmOut rpio.Pin
+	status rpio.Pin
 }
 
-func NewKeyer(config *Config) *Keyer {
-	keyer := Keyer{
+func NewPiGPIO(config *Config) *PiGPIO {
+	pigpio := PiGPIO{
 		config: config,
 		pwm:    false,
 	}
-	return &keyer
+	return &pigpio
 }
 
 // Set up inputs and outputs
-func (g *Keyer) Open() error {
+func (g *PiGPIO) Open() error {
 	err := rpio.Open()
 	if err != nil {
 		return err
@@ -76,8 +74,7 @@ func (g *Keyer) Open() error {
 	outLED := g.config.GPIOPins.SignalLED
 
 	// receiving morse from a GPIO
-	dahPin := g.config.GPIOPins.KeyLeft
-	ditPin := g.config.GPIOPins.KeyRight
+	inPin := g.config.GPIOPins.KeyLeft
 
 	statusPin := g.config.GPIOPins.StatusLED
 
@@ -87,12 +84,9 @@ func (g *Keyer) Open() error {
 	g.output.Low()
 
 	// Input pin
-	g.dahInput = rpio.Pin(dahPin)
-	g.dahInput.Input()
-	g.dahInput.PullUp()
-	g.ditInput = rpio.Pin(ditPin)
-	g.ditInput.Input()
-	g.ditInput.PullUp()
+	g.input = rpio.Pin(inPin)
+	g.input.Input()
+	g.input.PullUp()
 
 	// Status LED
 	g.status = rpio.Pin(statusPin)
@@ -102,30 +96,24 @@ func (g *Keyer) Open() error {
 	return nil
 }
 
-func (g *Keyer) Bit() bool {
+// ready Morse In hardware
+func (g *PiGPIO) Bit() bool {
+	if g.input.Read() == rpio.High {
+		return false
+	} else {
+		return true
+	}
+}
+
+func (g *PiGPIO) Dot() bool {
+	return false
+}
+func (g *PiGPIO) Dash() bool {
 	return false
 }
 
-// ready Morse In hardware
-func (g *Keyer) Dot() bool {
-	if g.ditInput.Read() == rpio.High {
-		return false
-	} else {
-		return true
-	}
-}
-
-// ready Morse In hardware
-func (g *Keyer) Dash() bool {
-	if g.dahInput.Read() == rpio.High {
-		return false
-	} else {
-		return true
-	}
-}
-
 // Set Morse Out hardware
-func (g *Keyer) SetBit(bit0 bool) {
+func (g *PiGPIO) SetBit(bit0 bool) {
 	if bit0 {
 		g.output.High()
 		g.SetToneOut(true)
@@ -136,7 +124,7 @@ func (g *Keyer) SetBit(bit0 bool) {
 }
 
 // Set PWM on/off
-func (g *Keyer) SetToneOut(v bool) {
+func (g *PiGPIO) SetToneOut(v bool) {
 	if g.pwm {
 		var dutyLen uint32
 
@@ -151,11 +139,11 @@ func (g *Keyer) SetToneOut(v bool) {
 }
 
 // Close the interface
-func (g *Keyer) Close() {
+func (g *PiGPIO) Close() {
 	g.status.Low()
 }
 
-func (g *Keyer) SetStatusLED(s bool) {
+func (g *PiGPIO) SetStatusLED(s bool) {
 	if s {
 		g.status.High()
 	} else {
